Fix 'implicit conversion from float to double' errors. (part 4)
These warnings showed up when enabling the '-Wdouble-promotion' GCC
compiler flag.  This batch of fixes consists of explicitly converting
from double to float.  These changes were made to match what seemed to
be the intent of the calculations.

+++ mythplugins/mythgallery/mythgallery/glsingleview.cpp
@@ -978,7 +978,7 @@ void GLSingleView::EffectFlutter(void)
                 m_effect_flutter_points[x][y][1] =
                     (float) (y / 20.0f - 1.0f) * ta.GetTextureY();
                 m_effect_flutter_points[x][y][2] =
-                    (float) sin((x / 20.0f - 1.0f) * M_PI * 2.0f) / 5.0;
+                    (float) sin((x / 20.0f - 1.0f) * static_cast<float>(M_PI) * 2.0f) / 5.0f;
             }
         }
     }
